converter exception? something wrong with json driverInfo
