+++ src/main/java/com/app/configuration/WebSecurityConfig.java
@@ -36,6 +36,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers("/driverInfo").hasRole("DRIVER")
                 .antMatchers("/driverLogin").hasRole("DRIVER")
                 .antMatchers("/selectProfile").hasRole("DRIVER")
+                .antMatchers("/driverPage").hasRole("DRIVER")
                 .antMatchers("/drivers").hasRole("MANAGER")
                 .antMatchers("/waggons").hasRole("MANAGER")
                 .antMatchers("/cargoes").hasRole("MANAGER")
@@ -43,7 +44,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers("/orders").hasRole("MANAGER")
                 .antMatchers("/points").hasRole("MANAGER")
                 .antMatchers("/cities").hasRole("MANAGER")
-                .anyRequest().hasAnyRole("ADMIN","MANAGER","DRIVER")
+                .anyRequest().hasAnyRole("ADMIN", "MANAGER", "DRIVER")
                 .and()
                 .authorizeRequests().antMatchers("/login**").permitAll()
                 .and()
