+++ src/main/java/com/app/controllers/LoginController.java
@@ -1,43 +1,78 @@
 package com.app.controllers;
 
 import com.app.DTO.DriverDTO;
-import com.app.model.Driver;
+import com.app.DTO.UserDriverDTO;
+import com.app.DTO.WayPointDTO;
+import com.app.model.UserDriver;
+import com.app.service.DriverLoginServiceInterface;
+import com.app.service.DriverServiceInterface;
 import com.app.service.UserDriverServiceInterface;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.ModelAndView;
 
-import java.security.Principal;
+import java.util.List;
 
 @Controller
+@RequestMapping(path = "/driverPage")
 public class LoginController {
-
     @Autowired
-    private UserDriverServiceInterface userDriverService;
+    DriverLoginServiceInterface driverLoginService;
+    @Autowired
+    DriverServiceInterface driverService;
+    @Autowired
+    UserDriverServiceInterface userDriverService;
 
-    @GetMapping("/")
-    public String index(Model model, Principal principal) {
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        int id = 0;
-        boolean hasUserRole = auth.getAuthorities().stream()
-                .anyMatch(r -> r.getAuthority().equals("ROLE_DRIVER"));
-        if (hasUserRole) {
-            DriverDTO driverDTO = userDriverService.findDriverIdByUsername(principal.getName());
-            id = driverDTO.getIdDriver();
-            if (id == -1) {
-                return "driverLogin";
-            } else {
-                model.addAttribute("message", "You are logged in as " + principal.getName());
-                model.addAttribute("driverID", id);
-                return "welcomeDriverPage";
-            }
-        }
-        model.addAttribute("message", "You are logged in as " + principal.getName());
-        return "welcome";
+    @RequestMapping(method = RequestMethod.GET)
+    public @ResponseBody
+    List<DriverDTO> getAllDriversJson() {
+        List<DriverDTO> drivers = driverService.getAllDriversJson();
+        return drivers;
+    }
+
+
+    @RequestMapping(path = "/{id}", method = RequestMethod.GET)
+    public ModelAndView driverInfoPage(@PathVariable int id) {
+        ModelAndView modelAndView = new ModelAndView("driverInfo");
+        return modelAndView;
+    }
+
+    @RequestMapping(path = "driverInfo/{id}", method = RequestMethod.GET)
+    public @ResponseBody
+    DriverDTO getDriverDTOByUser(@PathVariable int id) {
+        DriverDTO driverDTO = driverService.getDriverDTO(id);
+        return driverDTO;
+    }
+
+    @RequestMapping(path = "/companions", method = RequestMethod.GET)
+    public @ResponseBody
+    List<DriverDTO> loadCompanion(@PathVariable int id) {
+        List<DriverDTO> list = driverLoginService.findAllCompanions(id);
+        return list;
+    }
+
+    @RequestMapping(path = "/points", method = RequestMethod.GET)
+    public @ResponseBody
+    List<WayPointDTO> loadWayPoints(@PathVariable int id) {
+        List<WayPointDTO> points = driverLoginService.findAllOrderPoints(id);
+        return points;
+    }
+
+
+//
+//    @RequestMapping(value = "/{id}", method = RequestMethod.POST)
+//    public @ResponseBody
+//    DriverDTO editDriverUser(@RequestBody DriverDTO driverDTO, @PathVariable int id) {
+////        DriverDTO savedWaggon = driverService.saveChange(DriverDTO);
+//        return savedWaggon;
+//    }
+
+    @RequestMapping(method = RequestMethod.POST)
+    public @ResponseBody
+    UserDriver createUserDriver(@RequestBody UserDriverDTO userDriverDTO) {
+        System.out.println(userDriverDTO);
+        userDriverService.createUserDriver(userDriverDTO);
+        return new UserDriver();
     }
 }
