+++ src/main/java/com/app/repository/UserDriverRepositoryInterface.java
@@ -4,8 +4,6 @@ package com.app.repository;
 import com.app.model.Driver;
 import com.app.model.UserDriver;
 
-import java.util.Map;
-
 public interface UserDriverRepositoryInterface {
 
     UserDriver createUserDriver(UserDriver userDriver);
