+++ src/main/java/com/app/service/DriverLoginServiceInterface.java
@@ -2,8 +2,6 @@ package com.app.service;
 
 import com.app.DTO.DriverDTO;
 import com.app.DTO.WayPointDTO;
-import com.app.model.Driver;
-import com.app.model.WayPoint;
 
 import java.util.List;
 
@@ -14,7 +12,7 @@ public interface DriverLoginServiceInterface {
 
     int findOrderNumber(int idDriver);
 
-    List<WayPointDTO>  findAllOrderPoints(int idDriver);
+    List<WayPointDTO> findAllOrderPoints(int idDriver);
 
     int settingDriverTime(int idDriver);
