+++ src/main/java/com/app/service/impl/DriverLoginService.java
@@ -5,7 +5,6 @@ import com.app.DTO.WayPointDTO;
 import com.app.model.Driver;
 import com.app.model.WayPoint;
 import com.app.repository.DriverLoginRepositoryInterface;
-import com.app.repository.WayPointRepositoryInterface;
 import com.app.service.DriverLoginServiceInterface;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,21 +12,24 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+
 @Service
 public class DriverLoginService implements DriverLoginServiceInterface {
     @Autowired
     private DriverLoginRepositoryInterface driverLoginRepository;
+
     @Transactional(readOnly = true)
     @Override
     public List<DriverDTO> findAllCompanions(int idDriver) {
         List<Driver> listDriver = driverLoginRepository.findAllCompanions(idDriver);
         List<DriverDTO> listDriverDTO = new ArrayList<>();
-        for(Driver driver : listDriver){
+        for (Driver driver : listDriver) {
             DriverDTO driverDTO = new DriverDTO(driver);
             listDriverDTO.add(driverDTO);
         }
         return listDriverDTO;
     }
+
     @Transactional(readOnly = true)
     @Override
     public String findCurrentWaggon(int idDriver) {
@@ -43,7 +45,7 @@ public class DriverLoginService implements DriverLoginServiceInterface {
     public List<WayPointDTO> findAllOrderPoints(int idDriver) {
         List<WayPoint> wayPoints = driverLoginRepository.findAllOrderPoints(idDriver);
         List<WayPointDTO> wayPointsDTO = new ArrayList<>();
-        for(WayPoint wayPoint : wayPoints){
+        for (WayPoint wayPoint : wayPoints) {
             WayPointDTO pointDTO = new WayPointDTO(wayPoint);
             wayPointsDTO.add(pointDTO);
         }
