+++ src/main/webapp/WEB-INF/page/driverInfo.jsp
@@ -3,12 +3,12 @@
          pageEncoding="ISO-8859-1" %>
 <html>
 <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1"/>
     <script src="<c:url value="/resources/js/jQuery.js"/>"></script>
     <script src="<c:url value="/resources/js/driverInfo.js"/>"></script>
     <link rel="stylesheet" href="/resources/bootstrap/css/bootstrap.min.css">
     <link rel="stylesheet" href="/resources/bootstrap/css/bootstrap-grid.min.css">
     <link rel="stylesheet" href="/resources/bootstrap/css/bootstrap-reboot.min.css">
-    <meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1"/>
     <title>Driver Info</title>
 </head>
 <body>
